+++ scripts/3.1_model_view.py
@@ -1,22 +1,11 @@
-import os
-
-from compas_view2.app import App
-from compas_view2.collections import Collection
-
+from compas_viewer import Viewer
 from compas_ifc.model import Model
 
-HERE = os.path.dirname(__file__)
-FILE = os.path.join(
-    HERE,
-    "..",
-    "data",
-    "wall-with-opening-and-window.ifc",
-)
-model = Model(FILE)
-viewer = App(enable_sceneform=True)
+model = Model("data/wall-with-opening-and-window.ifc")
+
+viewer = Viewer()
 
 for entity in model.get_entities_by_type("IfcBuildingElement"):
-    print("Converting to brep:", entity)
-    obj = viewer.add(Collection(entity.body_with_opening), name=entity.name)
+    viewer.add(entity.body_with_opening, name=entity.name, **entity.style)
 
-viewer.run()
+viewer.show()