+++ scripts/4.1_edit_export.py
@@ -1,17 +1,7 @@
-import os
-
 from compas_ifc.model import Model
 
-HERE = os.path.dirname(__file__)
-FILE = os.path.join(
-    HERE,
-    "..",
-    "data",
-    "wall-with-opening-and-window.ifc",
-)
+model = Model("data/wall-with-opening-and-window.ifc")
 
-
-model = Model(FILE)
 
 print("\n" + "*" * 53)
 print("Export Examples")