+++ src/compas_ifc/__init__.py
@@ -28,7 +28,7 @@ __author__ = ["tom van mele"]
 __copyright__ = "ETH Zurich"
 __license__ = "MIT License"
 __email__ = "van.mele@arch.ethz.ch"
-__version__ = "0.2.0"
+__version__ = "0.3.0"
 
 
 HERE = os.path.dirname(__file__)
@@ -40,3 +40,7 @@ TEMP = os.path.abspath(os.path.join(HOME, "temp"))
 
 
 __all__ = ["HOME", "DATA", "DOCS", "TEMP"]
+
+__all_plugins__ = [
+    "compas_ifc.brep",
+]