+++ src/compas_ifc/model.py
@@ -13,6 +13,7 @@ from compas_ifc.entities.site import Site
 
 from .reader import IFCReader
 from .writer import IFCWriter
+import ifcopenshell
 
 
 class Model:
@@ -58,8 +59,8 @@ class Model:
 
     """
 
-    def __init__(self, filepath: str = None, entity_types: dict = None) -> None:
-        self.reader = IFCReader(model=self, entity_types=entity_types)
+    def __init__(self, filepath: str = None, entity_types: dict = None, use_occ=False, schema=None) -> None:
+        self.reader = IFCReader(model=self, entity_types=entity_types, use_occ=use_occ)
         self.writer = IFCWriter(model=self)
         self._new_entities = set()
         self._projects = None
@@ -69,6 +70,10 @@ class Model:
         self._elements = None
         self._building_elements = None
         self._geographic_elements = None
+        self._schema = None
+        if schema:
+            self._schema = ifcopenshell.ifcopenshell_wrapper.schema_by_name(schema)
+
         if filepath:
             self.open(filepath)
 
@@ -109,9 +114,22 @@ class Model:
         """Print the spatial hierarchy of the model."""
         self.project.print_spatial_hierarchy(max_level)
 
+    def print_summary(self):
+        """Print a summary of the model."""
+
+        print("=" * 80)
+        print("File: {}".format(self.reader.filepath))
+        print("Size: {} MB".format(self.reader.file_size()))
+        print("Project: {}".format(self.project.name))
+        print("Description: {}".format(self.project.attributes.get("Description", "")))
+        print("Number of sites: {}".format(len(self.sites)))
+        print("Number of buildings: {}".format(len(self.buildings)))
+        print("Number of building elements: {}".format(len(self.building_elements)))
+        print("=" * 80)
+
     @property
     def schema(self) -> str:
-        return self.reader._schema
+        return self._schema or self.reader._schema
 
     @property
     def project(self) -> Project: