+++ src/compas_ifc/resources/representation.py
@@ -5,10 +5,10 @@ from compas.geometry import Box
 from compas.geometry import Cone
 from compas.geometry import Cylinder
 from compas.geometry import Sphere
-from compas_occ.brep import BRep
 
 from .brep import brep_to_ifc_advanced_brep
 from .mesh import mesh_to_IfcPolygonalFaceSet
+from .mesh import mesh_to_IfcShellBasedSurfaceModel
 from .shapes import box_to_IfcBlock
 from .shapes import cone_to_IfcRightCircularCone
 from .shapes import cylinder_to_IfcRightCircularCylinder
@@ -16,6 +16,8 @@ from .shapes import sphere_to_IfcSphere
 
 
 def write_body_representation(file, body, ifc_entity, context):
+    from compas_occ.brep import OCCBrep
+
     def _body_to_shape(body):
         if isinstance(body, Box):
             shape = box_to_IfcBlock(file, body)
@@ -26,8 +28,11 @@ def write_body_representation(file, body, ifc_entity, context):
         elif isinstance(body, Cylinder):
             shape = cylinder_to_IfcRightCircularCylinder(file, body)
         elif isinstance(body, Mesh):
-            shape = mesh_to_IfcPolygonalFaceSet(file, body)
-        elif isinstance(body, BRep):
+            if file.schema == "IFC4" or file.schema == "IFC4x3":
+                shape = mesh_to_IfcPolygonalFaceSet(file, body)
+            else:
+                shape = mesh_to_IfcShellBasedSurfaceModel(file, body)
+        elif isinstance(body, OCCBrep):
             shape = brep_to_ifc_advanced_brep(file, body)
         else:
             raise Exception("Unsupported body type.")
@@ -46,11 +51,16 @@ def write_body_representation(file, body, ifc_entity, context):
         if not isinstance(shape, list):
             shape = [shape]
 
+    RepresentationType = "SolidModel"
+
+    if shape and shape[0].is_a("IfcShellBasedSurfaceModel"):
+        RepresentationType = "SurfaceModel"
+
     representation = file.create_entity(
         "IfcShapeRepresentation",
         ContextOfItems=context,
         RepresentationIdentifier="Body",
-        RepresentationType="SolidModel",
+        RepresentationType=RepresentationType,
         Items=shape,
     )